Advanced-Text-to-Image-Generation-main/requirements.txt
streamlit==1.39.0
requests==2.32.3
pillow==10.4.0
mlflow==2.9.1